requirements.txt
# Core LLM & Agent Libs
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
openai==1.77.0
tiktoken==0.9.0
pydantic==2.11.4 # Core dependency for LangChain/OpenAI

# Config & Utilities
PyYAML==6.0.2 # For prompts.yaml
python-dotenv==1.1.0 # For potential .env file usage
rich # For CLI output enhancement
